requirements.txt
flask
waitress
requests
pymupdf 
pillow
pandas
numpy
google-cloud-documentai
requests
